moving tablefunctions
I moved the tablefunctions out of fetch
